Add files via upload
AgriWebHub is a revolutionary online platform designed to bridge the gap between farmers and consumers, transforming the agricultural landscape into a more transparent and efficient ecosystem. With a mission to empower farmers while ensuring fair prices for consumers, AgriWebHub facilitates direct and seamless communication, eliminating the need for intermediaries and reducing the risk of fraud associated with traditional agricultural practices.